app/requirements.txt
debugpy
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
sqlalchemy